moved `into_robj` to their own module
