+++ extendr-api/src/robj/into_robj.rs
@@ -1,6 +1,8 @@
 use super::*;
 use crate::single_threaded;
 
+mod nonzero_into_robj;
+
 pub(crate) fn str_to_character(s: &str) -> SEXP {
     unsafe {
         if s.is_na() {
@@ -72,7 +74,7 @@ where
 }
 
 /// `ToVectorValue` is a trait that allows many different types
-/// to be converted to vectors. It is used as a type parameter
+/// to be converted to R vectors. It is used as a type parameter
 /// to `collect_robj()`.
 pub trait ToVectorValue {
     fn sexptype() -> SEXPTYPE {
@@ -376,29 +378,6 @@ impl ToVectorValue for Option<bool> {
         }
     }
 }
-
-macro_rules! impl_into_robj_nonzero {
-    ($type: ty) => {
-        impl From<$type> for Robj {
-            fn from(value: $type) -> Self {
-                // A nonzero Rust type is automatically guaranteed
-                // to be a valid R type if the normal type can be converted to R
-                value.get().into()
-            }
-        }
-    };
-}
-
-impl_into_robj_nonzero!(std::num::NonZeroU8);
-impl_into_robj_nonzero!(std::num::NonZeroU16);
-impl_into_robj_nonzero!(std::num::NonZeroU32);
-impl_into_robj_nonzero!(std::num::NonZeroU64);
-impl_into_robj_nonzero!(std::num::NonZeroUsize);
-impl_into_robj_nonzero!(std::num::NonZeroI8);
-impl_into_robj_nonzero!(std::num::NonZeroI16);
-impl_into_robj_nonzero!(std::num::NonZeroI32);
-impl_into_robj_nonzero!(std::num::NonZeroI64);
-// NOTE: `NonZeroIsize` is missing
 
 // Not thread safe.
 fn fixed_size_collect<I>(iter: I, len: usize) -> Robj
